LIST Activity

+++ app/src/main/java/com/example/hawk/imhungry/Restaurant.java
@@ -11,8 +11,8 @@ public class Restaurant {
     private double rating;
     private String address;
     private String phone;
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
 
     public String getName() {
         return name;
@@ -54,19 +54,19 @@ public class Restaurant {
         this.phone = phone;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
 }
